yeschef-backend/requirements.txt
fastapi
uvicorn
python-dotenv
postgrest
supabase-auth
google-genai
supadata
firecrawl-py
gunicorn
websockets
pydantic
python-multipart
livekit-agents>=1.0
livekit-plugins-google>=1.0
livekit-api>=1.0
